+++ source/extensions/omni.flux.layer_tree.usd.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.8.1]
+### Changed
+- Use generic centralized LayerTree model
+
 ## [1.8.0]
 ### Changed
 - Use centralized LayerTree widget