+++ source/extensions/omni.flux.property_widget_builder.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.6.1]
+### Changed
+- Use generic centralized LayerTree model
+
 ## [2.6.0]
 ### Changed
 - Use centralized LayerTree widget